chore: refactor token code to abstract processing code
correct typo in readme
